Use slightly more reliable way of getting user IDs
Before, we relied on parsing the profile picture ID; that doesn't work
when the user has the default profile picture.

Objectively the best thing to do would be to fetch the actual API for
this (/users/<username>), but I'm lazy. (It would mean passing more data
to parseProfile, which is a little yucky.)